Add `credentials_provider` crate
