+++ crates/credentials_provider/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "credentials_provider"
+version = "0.1.0"
+edition.workspace = true
+publish.workspace = true
+license = "GPL-3.0-or-later"
+
+[lints]
+workspace = true
+
+[lib]
+path = "src/credentials_provider.rs"
+
+[dependencies]
+anyhow.workspace = true
+fs.workspace = true
+gpui.workspace = true
