+++ crates/credentials_provider/src/credentials_provider.rs
@@ -1,0 +1,30 @@
+use std::future::Future;
+use std::pin::Pin;
+
+use anyhow::Result;
+use gpui::AsyncApp;
+
+/// A provider for credentials.
+///
+/// Used to abstract over reading and writing credentials to some form of
+/// persistence (like the system keychain).
+pub trait CredentialsProvider<T> {
+    /// Reads the credentials from the provider.
+    fn read_credentials<'a>(
+        &'a self,
+        cx: &'a AsyncApp,
+    ) -> Pin<Box<dyn Future<Output = Option<T>> + 'a>>;
+
+    /// Writes the credentials to the provider.
+    fn write_credentials<'a>(
+        &'a self,
+        credentials: T,
+        cx: &'a AsyncApp,
+    ) -> Pin<Box<dyn Future<Output = Result<()>> + 'a>>;
+
+    /// Deletes the credentials from the provider.
+    fn delete_credentials<'a>(
+        &'a self,
+        cx: &'a AsyncApp,
+    ) -> Pin<Box<dyn Future<Output = Result<()>> + 'a>>;
+}
